requirements.txt
asn1crypto==0.24.0
astroid==1.6.0
autopep8==1.3.5
backports.functools-lru-cache==1.5
configparser==3.5.0
cryptography==2.1.4
cycler==0.10.0
enum34==1.1.6
futures==3.2.0
http==0.2
idna==2.6
ipaddress==1.0.17
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
logilab-common==1.4.1
matplotlib==2.2.2
mccabe==0.6.1
NetfilterQueue==0.8.1
netifaces==0.10.7
numpy==1.14.3
psutil==5.4.2
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.8.3
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pyxdg==0.25
scapy==2.3.3
scapy-http==1.8
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
wrapt==1.9.0
